refactor(client): split scenarioClient.ts into packages/client/src/scenarios/
Moves the scenarios surface into its own folder, one symbol per file:

- ScenarioClient.ts: brand, ScenarioClient interface, NonScenarioClient
  type, and the shared buildScenarioClient internal helper.
- withScenario.ts: the sync attach-to-existing-scenario function.
- createScenario.ts: the async create-fresh-scenario function.

Test file split into withScenario.test.ts and createScenario.test.ts.
The unstable-do-not-use barrel re-exports from the new paths. Pure
re-organization — no behavior change.
