+++ packages/client/src/public/unstable-do-not-use.ts
@@ -25,8 +25,12 @@ export type { OsdkConfig } from "../public-utils/osdkConfig.js";
 
 export { createClientWithTransaction } from "../createClient.js";
 
-export { createScenario, withScenario } from "../scenarioClient.js";
-export type { NonScenarioClient, ScenarioClient } from "../scenarioClient.js";
+export { createScenario } from "../scenarios/createScenario.js";
+export type {
+  NonScenarioClient,
+  ScenarioClient,
+} from "../scenarios/ScenarioClient.js";
+export { withScenario } from "../scenarios/withScenario.js";
 
 export {
   applyShapeTransformations,
