--- packages/client/src/scenarioClient.ts
@@ -1,157 +1,0 @@
-/*
- * Copyright 2026 Palantir Technologies, Inc. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-import { OntologyScenarios } from "@osdk/foundry.ontologies";
-import { additionalContext, type Client } from "./Client.js";
-import { createClientWithScenario } from "./createClient.js";
-import type { MinimalClient } from "./MinimalClientContext.js";
-
-declare const scenarioBrand: unique symbol;
-
-/**
- * A {@link Client} attached to an ontology scenario. All read and write operations performed via this client are
- * scoped to that scenario. `ScenarioClient` is a superset of {@link Client}: it is usable anywhere a `Client` is
- * accepted (e.g. as the `scenarioReference` parameter on a scenario-merge action).
- *
- * @beta This is an experimental, unstable feature subject to change.
- */
-export interface ScenarioClient extends Client {
-  /** @internal */
-  readonly [scenarioBrand]: true;
-
-  /**
-   * Returns the scenario reference for this client. Pass directly to an action whose parameter is typed as
-   * `scenarioReference`, or use to inspect the underlying scenario RID.
-   */
-  getScenarioReference(): { scenarioRid: string };
-}
-
-/**
- * The set of clients that may be passed to {@link withScenario} or {@link createScenario}. Excludes already-branded
- * {@link ScenarioClient} instances at compile time to prevent nested-scenario anti-patterns.
- */
-export type NonScenarioClient = Client & {
-  readonly [scenarioBrand]?: never;
-};
-
-function buildScenarioClient(
-  parent: NonScenarioClient,
-  scenarioRid: string,
-): ScenarioClient {
-  const ctx: MinimalClient = parent[additionalContext];
-
-  if (ctx.transactionId != null) {
-    throw new Error(
-      "withScenario / createScenario: the supplied client already has an active transaction. Scenarios cannot be nested on transactions.",
-    );
-  }
-  if (ctx.scenarioRid != null) {
-    throw new Error(
-      "withScenario / createScenario: the supplied client already has an active scenario. Scenarios cannot be nested.",
-    );
-  }
-
-  const inner: Client = createClientWithScenario(
-    scenarioRid,
-    ctx.baseUrl,
-    ctx.ontologyRid,
-    ctx.tokenProvider,
-    {
-      logger: ctx.logger,
-      UNSTABLE_DO_NOT_USE_BRANCH: ctx.branch,
-    },
-    ctx.fetch,
-  );
-
-  return Object.defineProperties(inner, {
-    getScenarioReference: {
-      value: () => ({ scenarioRid }),
-    },
-  }) as ScenarioClient;
-}
-
-/**
- * Attach to an existing ontology scenario. Synchronous — no network call is made. The returned client scopes all
- * subsequent operations (`fetchPage`, `applyAction`, `executeFunction`, etc.) to the given `scenarioRid`.
- *
- * @param client - The base {@link Client} to derive context (`baseUrl`, `ontologyRid`, `tokenProvider`, `branch`, …)
- *   from. Must not already be a {@link ScenarioClient} or a transactional client.
- * @param scenarioRid - The RID of the scenario to attach to.
- * @returns a {@link ScenarioClient} bound to `scenarioRid`.
- *
- * @beta This is an experimental, unstable feature subject to change.
- *
- * @example
- * ```ts
- * import { withScenario } from "@osdk/client/unstable-do-not-use";
- *
- * const scenario = withScenario(client, "ri.actions..scenario.abc");
- * const page = await scenario(Doctor).fetchPage();
- * await scenario(Schedule).applyAction({ doctor: page.data[0], shift: "AM" });
- * ```
- */
-export function withScenario(
-  client: NonScenarioClient,
-  scenarioRid: string,
-): ScenarioClient {
-  return buildScenarioClient(client, scenarioRid);
-}
-
-/**
- * Mint a fresh ontology scenario and return a client scoped to it. Calls
- * `POST /v2/ontologies/{ontology}/scenarios/create` under the hood.
- *
- * @param client - The base {@link Client} to derive context (`baseUrl`, `ontologyRid`, `tokenProvider`, `branch`, …)
- *   from. Must not already be a {@link ScenarioClient} or a transactional client. When the base client has a branch
- *   set, the newly minted scenario uses that branch as its base.
- * @returns a {@link ScenarioClient} bound to the freshly minted scenario RID.
- *
- * @beta This is an experimental, unstable feature subject to change.
- *
- * @example
- * ```ts
- * import { createScenario } from "@osdk/client/unstable-do-not-use";
- *
- * const scenario = await createScenario(client);
- * const { scenarioRid } = scenario.getScenarioReference();
- * ```
- */
-export async function createScenario(
-  client: NonScenarioClient,
-): Promise<ScenarioClient> {
-  const ctx: MinimalClient = client[additionalContext];
-
-  if (ctx.transactionId != null) {
-    throw new Error(
-      "createScenario: the supplied client already has an active transaction. Scenarios cannot be nested on transactions.",
-    );
-  }
-  if (ctx.scenarioRid != null) {
-    throw new Error(
-      "createScenario: the supplied client already has an active scenario. Scenarios cannot be nested.",
-    );
-  }
-
-  const response = await OntologyScenarios.createScenario(
-    ctx,
-    await ctx.ontologyRid,
-    ctx.branch != null
-      ? { base: { type: "branch", branch: ctx.branch } }
-      : {},
-  );
-
-  return buildScenarioClient(client, response.scenarioRid);
-}
